+++ test/clj/matcher_combinators/midje_test.clj
@@ -1,7 +1,6 @@
 (ns matcher-combinators.midje-test
   (:require [midje.sweet :as midje :refer [fact facts => falsey]]
             [matcher-combinators.core :as core]
-            [matcher-combinators.dispatch :as dispatch]
             [matcher-combinators.matchers :as m]
             [matcher-combinators.midje :refer [match throws-match match-with match-roughly match-equals]]
             [matcher-combinators.model :as model]
@@ -296,42 +295,35 @@
                                ::result/value  (model/->Mismatch expected actual)
                                ::result/weight 1})))
 
-(def match-abs (match-with {java.lang.Long ->AbsValue}))
+(def match-abs (match-with [int? ->AbsValue]))
 
 (facts "match-with checker behavior"
   (core/indicates-match? (core/match -1 1)) => false
 
-  (fact "low-level invocation"
-    (core/indicates-match?
-     (dispatch/wrap-match-with
-      {java.lang.Long ->AbsValue}
-      (core/match -1 1)))
-    => true)
-
   (fact "using 2-arg match-with"
-    1 => (match-with {java.lang.Long ->AbsValue} -1)
-    -1 => (match-with {java.lang.Long ->AbsValue} 1))
+    1 => (match-with [int? ->AbsValue] -1)
+    -1 => (match-with [int? ->AbsValue] 1))
   (fact "binding 1-arg match-with to new checker"
     1 => (match-abs -1)
-    -1 => (match-abs 1)))
+    -1 => (match-abs 1))
+  (let [payload {:a {:b {:c 1}
+                     :d {:e {:inner-e {:x 1 :y 2}}
+                         :f 5
+                         :g 17}}}]
+    (fact "nested maps inside of an `embeds` of a match-equals are treated as equals"
+          payload
+          =not=> (match-equals {:a {:b {:c 1}
+                                    :d (m/embeds {:e {:inner-e {:x 1}}})}})
+          payload
+          => (match-equals {:a {:b {:c 1}
+                                :d (m/embeds {:e {:inner-e {:x 1 :y 2}}})}}))))
 
 (facts "match-equals"
   (fact "normal loose matching passes"
     {:a 1 :b 3 :c 1} => (match {:a 1 :b odd?}))
   (fact "match-equals is more strict"
     {:a 1 :b 3 :c 1} =not=> (match-equals {:a 1 :b odd?})
-    {:a 1 :b 3} => (match-equals {:a 1 :b odd?}))
-  (let [payload {:a {:b {:c 1}
-                     :d {:e {:inner-e {:x 1 :y 2}}
-                         :f 5
-                         :g 17}}}]
-    (fact "nested maps inside of an `embeds` of a match-equals are treated as equals"
-      payload
-      =not=> (match-equals {:a {:b {:c 1}
-                                :d (m/embeds {:e {:inner-e {:x 1}}})}})
-      payload
-      => (match-equals {:a {:b {:c 1}
-                            :d (m/embeds {:e {:inner-e {:x 1 :y 2}}})}}))))
+    {:a 1 :b 3} => (match-equals {:a 1 :b odd?})))
 
 (fact "match-roughly"
   {:a 1 :b 3.05} => (match-roughly 0.1
@@ -340,7 +332,7 @@
                                        {:a 1 :b 3.0}))
 
 (fact "example from docstring"
-  5 => (match-with {java.lang.Long greater-than-matcher}
+  5 => (match-with [int? greater-than-matcher]
                    4))
 
 (spec.test/unstrument)